requirements.txt
pillow
python-magic
python-magic-bin
markdown
PythonVideoConverter